+++ jdbc/src/main/java/org/springframework/dao/IncorrectResultSizeDataAccessException.java
@@ -1,6 +1,6 @@
-package org.springframework.jdbc;
+package org.springframework.dao;
 
-public class IncorrectResultSizeDataAccessException extends RuntimeException {
+public class IncorrectResultSizeDataAccessException extends DataAccessException {
 
     public IncorrectResultSizeDataAccessException(int size) {
         super("Incorrect Result Size Data returned! returned data size was " + size);
