--- jdbc/src/main/java/org/springframework/jdbc/CanNotCloseResultSetException.java
@@ -1,8 +1,0 @@
-package org.springframework.jdbc;
-
-public class CanNotCloseResultSetException extends RuntimeException {
-
-    public CanNotCloseResultSetException(String msg) {
-        super(msg);
-    }
-}
