+++ jdbc/src/main/java/org/springframework/jdbc/datasource/DataSourceUtils.java
@@ -3,7 +3,7 @@ package org.springframework.jdbc.datasource;
 import java.sql.Connection;
 import java.sql.SQLException;
 import javax.sql.DataSource;
-import org.springframework.jdbc.CannotGetJdbcConnectionException;
+import org.springframework.dao.CannotGetJdbcConnectionException;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
 
 // 4단계 미션에서 사용할 것
